+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Features
+- The weather module now comes with default values for `temp_unit` and
+  `language`, so there is no need to specify them every time (unless those
+  defaults doesn't suit you).
+
 ## [v1.1.0] - 2020-04-02
 
 ### Features